integration: move public tests into file
- conventional approach with ginkgo
- table test not needed for the single config
- can't use the debugger with many nested tables
